create status form
